Legger skjemaet for et barn i egen komponent, slik at vi kan sjekke om all data finnes og eventuell redirecte hvis ikke før useSkjema-hooken blir kalt
